fix: normalize homework relation identifiers

+++ EduOS.Core/Entities/LMS/Homework.cs
@@ -6,10 +6,10 @@ namespace EduOS.Core.Entities.LMS
 {
     public class Homework : BaseTenantEntity
     {
-        public int ClassId { get; set; }
-        public int SectionId { get; set; }
-        public int SubjectId { get; set; }
-        public int TeacherId { get; set; }
+        public long ClassId { get; set; }
+        public long SectionId { get; set; }
+        public long SubjectId { get; set; }
+        public long TeacherId { get; set; }
         public string Title { get; set; } = string.Empty;
         public string? Description { get; set; }
         public DateTime AssignedDate { get; set; }
